migrate work allocations into its own controller
